refactor: email-integratin compnent has now been chunkced down into subcomponetns and those subcomponents have been moved into a sub directory
