+++ frontend/src/components/email/components/HowItWorks.tsx
